# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"LogInResponse": {
				"properties": {
					"token": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"avatar"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Message": {
				"properties": {
					"from": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"idConversation": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"from",
					"to",
					"message",
					"idConversation",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Conversation": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"from": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"messages": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Message"
							},
							"type": "array"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"photo": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"to",
					"from",
					"date",
					"messages",
					"name",
					"photo"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "componentes",
		"version": "1.0.0",
		"description": "Backend for a real time chat application",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Jocselyn Aguilar"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/login": {
			"get": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LogInResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "password",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/conversations": {
			"get": {
				"operationId": "GetConversationsByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Conversation"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/message": {
			"post": {
				"operationId": "CreateMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Message"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Message"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}